+++ resources/js/app.js
@@ -25,9 +25,6 @@ if (apiToken) {
 window.Vue = require('vue');
 Vue.mixin(require('./i18n'));
 
-import VueMarkdown from 'vue-markdown'
-Vue.component('vue-markdown', VueMarkdown);
-
 import Toasted from 'vue-toasted';
 Vue.use(Toasted, {
     position: 'top-center',